refact: resolver commun to whole app; warp in mutex
